Extracted CLI logic to a module. Can now be re-used in tests & reduce mocks required
